respond to new NetworkManager "connectivity-change" event
Similar to https://phabricator.endlessm.com/T12336, move our if-up.d script to
NetworkManager/dispatcher.d so that we can respond to the connectivity-change
action which triggers correctly eg when the user has completed a captive
portal. Otherwise we can "miss" the internet being available because we
activate our script before the user has authenticated.