add activated file condition to systemd unit
On first boot the .timer job is activated and will inevitably run until the system reboots and the condition stops the timer from running. This means the python activate script is run a few times needlessly and then quits. In the case that the timer is going off but the script has run (eg after the network connectivity changing) then this makes systemd check that /var/lib/eos-phone-home/activated doesn't exist before ever starting the activate script, saving the overhead of the python intepreter.