T13240: fix broken systemd-run invocation
The systemd-run invocation in the NM dispatcher hook failed because it was trying to create a transient unit named eos-phone-home.timer which is already defined. Change it to create a transient unit which actually in turn calls systemctl to start eos-phone-home.service. This seems a little silly but also has the benefit that systemd will serialise invocations so the script will only be run once at a time.

https://phabricator.endlessm.com/T13240