Wait until FBE has completed
We would like to only send the activation message and pings once the FBE
has completed. This increases the chance that we have a network
connection (since setting up wifi is one step of the FBE), and allows us
to include accurate information as to whether metrics are enabled from
the very first ping. To achieve this, we:

* make the service conditional on /home being non-empty. The FBE only
  creates the new user once all other pages are complete.
* add a path unit which activates the timer when /home is non-empty.
* fire the timer when it is activated, and every 3h after that.

The net result is that the timer will be activated at the end of the FBE
and on every subsequent boot.