Don't create 'count' file until first ping is sent
We'd expect to see approximately the same number of activations as ping
0s submitted to the server, but in fact we see around 10% fewer ping 0s
than activations.

/var/lib/eos-phone-home/count is meant to contain the number of previous
successful pings. At most every 24 hours, we submit a ping containing
the contents of this file, then increment the file. Logically, if the
file does not exist (as is the case on new installations) the count of
previous pings is 0.

Previously, running this script without an active network
connection causes /var/lib/eos-phone-home/count to be written with value
'0'. The value is correct -- no pings have been sent yet -- but the
timestamp on this file is used to determine when to send the next ping.
If the file was written within the past 24 hours, no ping is sent. So,
if the network comes up within the next 24 hours, the activation message
will be sent, but not ping 0. After 24 hours, ping 0 would be sent, but
short-lived installations would not get that far.

Instead, only write /var/lib/eos-phone-home/count after a ping has been
sent (or in the time-goes-backwards error case to force the next ping to
happen in >= 24 hours). (This already worked correctly for pings after
the 0th.)

https://phabricator.endlessm.com/T16966