Don't exit failure on network errors
This script is run as a systemd service, and exiting non-0 causes
systemd to enter degraded mode, which leads to visible messages on
shutdown. But failure to contact the server is normal and expected in
many circumstances beyond the control of the local machine (such as when
you're offline, or when the server is undergoing maintenance).

Instead, by default, only exit non-zero if something local is
catastrophically wrong (such as the state directory being missing).

https://phabricator.endlessm.com/T18099