eos-phone-home: only provide "unknown" filler for mandatory strings
These are the keys where the server would reject the string for being
missing. All of the other get functions return a typed value (boolean
or integer) so don't need any default/filler value for mandatory
values.

The astute will note that "serial" is not in fact mandatory on the
server, but we implement the "unknown" filler for backwards compatibility
and to avoid creating potential duplicated/aliased records on the server.