eos-phone-home: calculate and submit mac_hash in activation messages
In order to tell apart systems where the serial number is not set
by the vendor, we add a mac_hash to the activate message. This is
an ADLER32 hash of the (lexicographically) first MAC address of the
non-removable ethernet interfaces, or if there are none, the wireless
interfaces.

The intention is to provide a value which is most likely to be stable
and unique for a given machine. Ethernet adapters, if present, are
more likely to be supported out of the box, whereas support for
wireless adapters could be added in a later OS release and would
cause a new value to be calculated on the same machine.

https://phabricator.endlessm.com/T20790