Add a test suite
"eos-phone-home" is not a valid Python module name, so we can't just
import it, but we won't let that stop us. We also need a way to make
PhoneHome read files rooted in a directory controlled by the test suite;
the approach taken here is to pass a string (defaulting to '') into its
constructor and prepend all paths with that string.

This test suite does not provide full coverage, but would at least have
caught https://phabricator.endlessm.com/T20993.